+++ CHANGES.rst
@@ -21,6 +21,7 @@ Unreleased
     2.56e-3. :issue:`912`, :pr:`922`
 -   Int and float literals can be written with the '_' separator for
     legibility, like 12_345. :pr:`923`
+-   Fix a bug causing deadlocks in ``LRUCache.setdefault``. :pr:`1000`
 
 
 Version 2.10.2